Make install script do the right things.
Previously it would do some wonky shit with the directories that were
being linked. This is now no longer a problem, as there is now actually
logic inside of the symlink loop.

Also check that zsh is installed before trying to change the shell, and
require that to be installed before moving on to anything else.

+++ install.sh
@@ -1,29 +1,47 @@
-DOTSTUFFDIR=$PWD
-
-if [ -d "$HOME/.oh-my-zsh" ]; then
-    rm -rf $HOME/.oh-my-zsh
+# make sure zsh is installed before continuing
+zsh=`command -v zsh`
+if [ "${zsh}x" = "x" ]; then
+	echo "Install zsh before continuing"
+	exit 1
 fi
 
-# overwrite anyways.
-for item in zshrc zsh_lib/ vim/ vimrc gvimrc gitconfig tmux.conf; do
-	#rm -rf ~/.$item
-	ln -s -f $DOTSTUFFDIR/$item ~/.$item
-done
-
-# install vim plugins
-vim +PlugUpgrade +PlugClean +PlugUpdate +PlugInstall +qall
-
-zsh=`which zsh`
+# set zsh to your shell
 if [ $SHELL != $zsh ]; then
+	# macos only allows you to change shells listed in /etc/shells
 	if [[ `uname -s` == "Darwin" && ! `grep $zsh /etc/shells` ]]; then
 		echo "Adding $zsh to /etc/shells"
 		echo "\n$zsh" | sudo tee -a /etc/shells > /dev/null
 	fi
 
-    echo "chsh"
-    chsh -s `which zsh`
+    chsh -s $zsh
 fi
 
-if [ -f ~/.aliases ]; then
-	mv ~/.aliases ~/.local_profile
+DOTSTUFFDIR=$PWD
+
+# remove oh-my-zsh if it's installed
+if [ -d "$HOME/.oh-my-zsh" ]; then
+    rm -rf $HOME/.oh-my-zsh
 fi
+
+# link items in home so they're, y'know, actually used
+for item in zshrc zsh_lib vim vimrc gvimrc gitconfig tmux.conf; do
+	dest=~/.$item
+	source=$DOTSTUFFDIR/$item
+
+	if [ -h $dest ]; then
+		if [ ! `readlink $dest` -ef $source ]; then
+			rm -f $dest
+			ln -s -f $source $dest
+		fi
+	else
+		if [ -d $dest ]; then
+			rm -rf $dest
+		else
+			rm -f $dest
+		fi
+		ln -s -f $source $dest
+	fi
+done
+
+# install vim plugins
+vim +PlugUpgrade +PlugClean +PlugUpdate +PlugInstall +qall
